+++ .gitignore
@@ -15,3 +15,4 @@ example
 vendor
 .ruby-version
 .bundle/
+.env
